refactor(ui): move components/data_table to dbflux_components (S4a) (#102)
