+++ crates/dbflux_components/src/components/data_table/events.rs
