+++ crates/dbflux_components/src/components/data_table/model.rs
