+++ crates/dbflux_components/src/components/mod.rs
@@ -1,0 +1,4 @@
+// Component modules relocated from dbflux_ui.
+// This file grows one line per sub-slice as each component is moved.
+
+pub mod data_table;
